microservices/generators/requirements.txt
boto3>=1.30.0
botocore>=1.30
pydantic>=2.0.0
kafka-python>=2.0.0
pymongo>=2.5.0
pika>=1.2.0